moved states and API calls into App.js instead of Board.js
